Added Classes, Button Functionality and Toasts
JavaDoc explains all of the methods in each class. You can now click on a name and it takes you to their profile, made name appear in the view user screen. New hints, toasts. A lot of these methods will be added onto/changed once we get the database going.